Implement an end-to-end prototype (#43)
* Create new page-grouped DB target, refactor accordingly

* Hook into manager reopen/load code

* Expose pages read from disk

* Extract persistence functionality into a separate function

* Implement the read, write, and scan paths

* Add sanity check PGDB tests

* Move page grouping tests into the main test executable

* Add PGLLSM as an option in run_custom

* Update YCSB run/preload scripts to include page-grouped LLSM

* Add updated initial YCSB experiments

* Script fixes

* Shuffle path fix

* Use 1 M requests

* Add missing workload config

* Fix compare exchange

* Increase back to 10 M

+++ scripts/e2e_custom/run.sh
@@ -86,7 +86,9 @@ code=$?
 kill -s SIGINT -- $iostat_pid
 wait
 
-cp $DB_PATH/$db_type/LOG $COND_OUT/$db_type.log
+if [ -e $DB_PATH/$db_type/LOG ]; then
+  cp $DB_PATH/$db_type/LOG $COND_OUT/$db_type.log
+fi
 du -b $DB_PATH >$COND_OUT/db_space.log
 
 # Report that the experiment failed if the `run_custom` exit code is not 0
